编码

+++ src/main/java/top/testops/others/classification/CommissionedClassification.java
@@ -22,7 +22,7 @@ public class CommissionedClassification implements PaymentClassification {
     }
 
     @Override
-    public int calculatePay(Timestamp date) {
+    public double calculatePay(Timestamp date) {
         return 0;
     }
 }
