+++ src/main/java/top/testops/others/classification/SalariedClassification.java
@@ -15,7 +15,7 @@ public class SalariedClassification implements PaymentClassification {
     }
 
     @Override
-    public int calculatePay(Timestamp date) {
+    public double calculatePay(Timestamp date) {
         return 0;
     }
 }
